Move service_unavailable, strip_infura_project_id to error_responses.py
